Another fix for the split system and realistic collision
- if collisonTime==particle.age-1 then it means that the particle has
collided with the specific surface in the last update cycle, when
realistic collision is on then the speed of that component will be 0.
This also fixes the split issue, where e.g. particles lying on the
ground colliding on the x or z side would gain speed (cf. video in
discord)